server/requirements.txt
flask
flask-cors
openai
tiktoken
beautifulsoup4
pytesseract
pillow
